Refactored generate_report method to reduce cognitive complexity
